feat: add electron-log file logging

+++ src/lib/ipfs-caching.js
@@ -2,6 +2,7 @@ const url = require('url')
 const { promisify } = require('util')
 const { IpfsConnector } = require('@akashaproject/ipfs-connector')
 const { session } = require('electron')
+const log = require('electron-log')
 const storage = require('./storage')
 
 // 7 days is the default expiration time if the pinned hash is not accessed
@@ -19,18 +20,21 @@ async function pinAragonClientForNetwork (newHash, network) {
   const storageKey = `${network}:aragon.aragonpm.eth`
   const storedHash = await storage.get(storageKey)
   if (storedHash !== newHash) {
+    log.debug(`Pinning new client hash for ${network}: ${newHash}`)
     await ipfsPinAdd(newHash)
     if (storedHash) {
+      log.debug(`Removing old client hash for ${network}: ${storedHash}`)
       await ipfsPinRm(storedHash)
     }
     await storage.set(storageKey, { hash: newHash })
-    console.log(`Pinned new client hash for ${network}:`, newHash)
+    log.info(`Pinned new client hash for ${network}: ${newHash}`)
   }
 }
 
 async function updateExpiration (hash) {
   const expirationDate = new Date()
   expirationDate.setDate((expirationDate.getDate() + IPFS_EXPIRATION))
+  log.debug(`Updating expiration date for ${hash}: ${expirationDate}`)
   await storage.set(hash, { expiration: expirationDate.getTime() })
 }
 
@@ -38,14 +42,13 @@ async function purgeUnusedIpfsResources () {
   const keys = await storage.keys()
   for (const key of keys) {
     const data = await storage.get(key)
-    if (data.expiration) {
-      if (data.expiration < new Date().getTime()) {
-        await ipfsPinRm(hash)
-        await storage.delete(hash)
-      }
+    if (data.expiration && data.expiration < new Date().getTime()) {
+      log.debug(`Purging unused IPFS resource: ${hash}`)
+      await ipfsPinRm(hash)
+      await storage.delete(hash)
     }
   }
-  console.log('Purged old IPFS resources')
+  log.info('Purged old IPFS resources')
 }
 
 function listenAndPinResources () {
@@ -56,8 +59,8 @@ function listenAndPinResources () {
       if (!(await storage.has(hash))) {
         await ipfsPinAdd(hash)
       }
-      updateExpiration(hash)
-      console.log('Pinned new IPFS resource:', hash)
+      await updateExpiration(hash)
+      log.info('Pinned new IPFS resource:', hash)
     }
     cb({ cancel: false, requestHeaders: details.requestHeaders })
   })
